Extract logic to hook
